Remove fs-extra dependency (#1891)
@zachleat proposed removing fs-extra since the build in FS module now provides a promise based option and supports recursive options.

This Commit only uses the ```require("fs")``` instead of ```require("fs").promises``` or ```require("fs/promises")```, since the last one is not support in Node Version <14 and the sync API is used.

Except for Template.js no file logic needed to be changed to remove fs-extra.

According to my run of "npm test" test results are not changed by this commit.

Signed-off-by: Raphael Höser <raphael@hoeser.info>