Fix #1242 Dot-Pathing
Dots in the global datafile path leads to data beeing incorrectly nested.
This commits changes the behavior so a datafile a.b.json will create an entry with data["a.b"] instead of data.a.b.
The old behavior was a bug according to #1242.

The implementation makes use of lodashs set and get functions which also take object paths in array form instead of string. That way we can implement this safely.

Signed-off-by: Raphael Höser <raphael@hoeser.info>