#1312 Improve situation for dependency monitoring
This change adds a "PathNormalizer" with tests which converts all paths to be unix compatible (replace "\" with "/"). That way auto reloading dependencies of the config also works on windows.

Sadly nested dependencies still always lag exactly one update behind, so if you have the following dependency graph:
- .eleventy.js depends on
- middle.js depends on
- nested.js

and you update nested, you see the following:
- do update 1 -> no result
- do update 2 -> see update 1
- do update 3 -> see update 2

During a short lookaround, I was unable to resolve this issue.

Signed-off-by: Raphael Höser <raphael@hoeser.info>