#1175 and #2287 Allow eleventyExcludeFromCollections to take a list of tags
This is a first implementation for #1175 and #2287 to gather feedback.

This change allows for more fine grained control over how a template is excluded from collections by allowing to set eleventyExcludeFromCollections to a tag or list of tags to exclude the template from.

This is meant as a first implementation for feedback. Also I'm not completely sure if excluding the `all` Collection is done correctly.

I only added minimal tests, but probably some more are needed.

Signed-off-by: Raphael Höser <raphael@hoeser.info>