Register dependencies found during pug compilation
My main layout is written in pug, and uses an `include` file. When
running `eleventy --serve` I noticed that changes to the included file
would trigger rebuilds, but not show the changes. This seems to be
similar to #2741.

Here I fix this by using the list of dependencies that the pug `compile`
function includes on the function that it returns. With this I am seeing
changes to my included file reflected immediately.